css changes, attempt to implement MixItUp
MixItUp doesn't work well with the cart.addItem method when an item is
clicked, something within the function is causing an item to be removed
and added back or something, I'm not entirely sure.

I also could not get it to filter properly on load. The filter buttons
will light up accordingly, but something is not working right with
angularJS's method of div generation. Maybe it's the order of execution.
Examples of MixItUp I've seen filters divs immediately on load.

CSS changes: implemented some colors and formatting stuff